renames the package from app to beesite
