+++ src/beesite/static/img/maps/box_small.png
