+++ src/beesite/static/img/maps/delta_small.png
