+++ src/beesite/static/img/maps/donut_small.png
