+++ src/beesite/static/img/maps/kilo_small.png
