+++ src/beesite/static/img/maps/pubby_small.png
