+++ src/beesite/static/js/bans.js
