+++ src/beesite/static/js/book.js
