+++ src/beesite/static/js/stats.js
