+++ src/beesite/templates/book.html
