+++ src/beesite/templates/stats.html
