+++ src/wsgi.py
@@ -1,4 +1,4 @@
-from app import app
+from beesite import app
 
 if __name__ == "__main__":
     app.run(host="localhost", port=8080, debug=True)
